react-django-mongo-app/backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
pymongo==4.3.3
python-dotenv==1.0.0
djongo==1.3.6
corsheaders==3.13.0
djangorestframework-simplejwt==5.2.2